completed

+++ src/main/java/ru/netology/service/PostService.java
@@ -17,7 +17,7 @@ public class PostService {
     return repository.all();
   }
 
-  public Post getById(long id) {
+  public Post getById(int id) {
     return repository.getById(id).orElseThrow(NotFoundException::new);
   }
 
@@ -25,7 +25,7 @@ public class PostService {
     return repository.save(post);
   }
 
-  public void removeById(long id) {
+  public void removeById(int id) {
     repository.removeById(id);
   }
 }
